feat: no document with staticExport

+++ packages/plugin-rax-app/src/setDev.js
@@ -79,7 +79,9 @@ module.exports = function (api) {
     // Remove outputDir when start devServer
     const { outputDir = 'build' } = userConfig;
     configs.forEach((config) => {
-      fs.emptyDirSync(path.resolve(rootDir, outputDir, config.name));
+      if (config.name) {
+        fs.emptyDirSync(path.resolve(rootDir, outputDir, config.name));
+      }
     });
 
     logWebpackConfig(configs);
